+++ projects/fusion-ui/components/dropdown-select/v4/dropdown-select-v4.component.ts
@@ -1,4 +1,4 @@
-import {ChangeDetectionStrategy, Component, Injector, Input} from '@angular/core';
+import {ChangeDetectionStrategy, Component, inject, Input} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {TooltipModule} from '@ironsource/fusion-ui/components/tooltip';
 import {IconModule} from '@ironsource/fusion-ui/components/icon/v1';
@@ -28,8 +28,10 @@ export class DropdownSelectV4Component {
     @Input() iconColor: string;
     @Input() testId: string;
     @Input() country: CountryCode | string;
-    testIdDropdownModifiers: typeof DropdownTestIdModifiers = DropdownTestIdModifiers;
-    testIdsService: TestIdsService = this.injector.get(TestIdsService);
+    @Input() hideCaretIcon: boolean = false;
 
-    constructor(private injector: Injector) {}
+    /** @internal */
+    testIdDropdownModifiers: typeof DropdownTestIdModifiers = DropdownTestIdModifiers;
+    /** @internal */
+    testIdsService: TestIdsService = inject(TestIdsService);
 }